refactor: split command up into subcommands
