+++ src/utils/error.ts
@@ -1,0 +1,2 @@
+/** Represents an error that will be displayed to the user by the client. */
+export class UserError extends Error {}
