+++ src/utils/index.ts
@@ -1,2 +1,4 @@
 export * from "./drizzle.ts";
 export * from "./common.ts";
+export * from "./discordjs.ts";
+export * from "./error.ts";
